requirements.txt
folium==0.17.0
pandas==2.2.3
plotly==5.24.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.38.0
streamlit_folium==0.23.0
